+++ paver/misctasks.py
@@ -1,21 +1,21 @@
 """Miscellaneous tasks that don't fit into one of the other groupings."""
-import os
 import pkgutil
 import zipfile
 from StringIO import StringIO
-
+from os.path import join, dirname, exists, abspath
 from paver.easy import dry, task
+from paver.tasks import VERSION, cmdopts
 
-_docsdir = os.path.join(os.path.dirname(__file__), "docs")
-if os.path.exists(_docsdir):
+_docsdir = join(dirname(__file__), "docs")
+if exists(_docsdir):
     @task
     def paverdocs():
         """Open your web browser and display Paver's documentation."""
         import webbrowser
-        webbrowser.open("file://" 
-            + (os.path.join(os.path.abspath(_docsdir), 'index.html')))
+        webbrowser.open("file://%s"  % join(abspath(_docsdir), 'index.html') )
         
 @task
+@cmdopts([('versioned_name', 'v', 'Determine if minilib uses version in its name')])
 def minilib(options):
     """Create a Paver mini library that contains enough for a simple
     pavement.py to be installed using a generated setup.py. This
@@ -23,6 +23,13 @@ def minilib(options):
     The output file is 'paver-minilib.zip' in the current directory.
     
     Options:
+
+    versioned_name
+        if set to True, paver version will be added into minilib's filename
+        (ie paver-minilib-1.1.0.zip)
+        purpose is to avoid import error while using different versions of minilib
+        with easy_install
+        (default False)
     
     extra_files
         list of other paver modules to include (don't include the .py 
@@ -33,16 +40,17 @@ def minilib(options):
     filelist = ['__init__', 'defaults', 'path', 'path25', 'release',
                 'setuputils', "misctasks", "options", "tasks", "easy"]
     filelist.extend(options.get('extra_files', []))
-    output_file = 'paver-minilib.zip'
+
+    output_version = "-%s" % VERSION if options.versioned_name else ""
+    output_file = 'paver-minilib%s.zip' % output_version
 
     def generate_zip():
         # Write the mini library to a buffer.
         buf = StringIO()
         destfile = zipfile.ZipFile(buf, "w", zipfile.ZIP_DEFLATED)
         for filename in filelist:
-            destfile.writestr(
-                "paver/" + (filename + ".py"),
-                pkgutil.get_data('paver', filename + ".py"))
+            destfile.writestr("paver/%s.py" % filename,
+                              pkgutil.get_data('paver', "%s.py" % filename))
         destfile.close()
 
         # Write the buffer to disk.
@@ -52,26 +60,32 @@ def minilib(options):
     dry("Generate %s" % output_file, generate_zip)
     
 @task
-def generate_setup():
+@cmdopts([('versioned_name', 'v', 'Determine if setup refer to minilib with version in its name')])
+def generate_setup(options):
     """Generates a setup.py file that uses paver behind the scenes. This 
     setup.py file will look in the directory that the user is running it
     in for a paver-minilib.zip and will add that to sys.path if available.
     Otherwise, it will just assume that paver is available."""
-    from paver.easy import dry
+    if options.versioned_name:
+        minilib_name = "paver-minilib%s.zip" % ("-" + VERSION)
+        is_versioned_msg = ', referring versioned minilib: %s' % minilib_name
+    else:
+        is_versioned_msg = ""
+        minilib_name = 'paver-minilib.zip'
+
     def write_setup():
         setup = open("setup.py", "w")
         setup.write("""try:
     import paver.tasks
 except ImportError:
-    import os
-    if os.path.exists("paver-minilib.zip"):
+    from os.path import exists
+    if exists("%(minilib_name)s"):
         import sys
-        sys.path.insert(0, "paver-minilib.zip")
+        sys.path.insert(0, "%(minilib_name)s")
     import paver.tasks
 
 paver.tasks.main()
-""")
+""" % {'minilib_name': minilib_name})
         setup.close()
         
-    dry("Write setup.py", write_setup)
-    
+    dry("Write setup.py%s" % is_versioned_msg, write_setup)