+++ docs/source/changelog.rst
@@ -12,6 +12,10 @@ Paver Changelog
 * ``@no_help`` to provide a way to hide task from ``paver help`` (github issue `#36 <https://github.com/paver/paver/issues/36>`_)
 * ``@might_call`` for more complex dependencies (see docs, not only github issue `#16 <https://github.com/paver/paver/issues/16>`_)
 * bundled path.py upgraded to patched version 2.2.2 (github issue `#15 <https://github.com/paver/paver/issues/15>`_)
+* correctly handle dependencies in ``install_requires`` directive for `setup.py install` command (github issue `#49 <https://github.com/paver/paver/issues/49>`_)
+* fix creating virtualenv (github issue `#44 <https://github.com/paver/paver/issues/44>`_)
+* fix virtualenv example in docs (github issue `#48 <https://github.com/paver/paver/issues/48>`_)
+* path.rename() do not call rename twice (github issue `#47 <https://github.com/paver/paver/issues/47>`_)
 
 1.0.5 (October 21, 2011)
 ------------------------